Added fragment section
The "Fragments" folder is used to store the fragment classes. Fragments are used to represent reusable UI portions or behaviors within an activity.
Fragments are like modular sections of an activity that can be combined or reused in different activities. They have their own lifecycle and can be added or removed dynamically to create flexible and adaptive UI designs.